+++ shorebird_code_push/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.0.5
+
+- docs: fix Discord badge in `README.md`
+
+# 2.0.4
+
+- feat: allow for custom UpdateTrack names in addition to stable, staging, and
+  beta
+
 # 2.0.3
 
 - feat: override `toString` in `ReadPatchException` and `UpdateException`